tools/setup-env: default to disable multimedia settings
imx-gpu-viv does not provide everything it needs to for virtual/libgl
on x11 backend or on Wayland backend with XWayland support. It depends
on mesa to fill in what is missing.

So when x11 is enabled in DISTRO_FEATURES, imx-gpu-viv depends on mesa,
but mesa is skipped as PREFERRED_PROVIDER_virtual/libgl set to imx-gpu-viv.

This causes below parse error:
| ERROR: Nothing PROVIDES 'mesa' (but imx-gpu-viv_6.4.0.p2.0-aarch64.bb DEPENDS on or otherwise requires it)
| mesa was skipped: PREFERRED_PROVIDER_virtual/libgl set to imx-gpu-viv, not mesa

Default to disable these settings only apply for fsl-image-multimedia-full.

Signed-off-by: Ting Liu <ting.liu@nxp.com>